Auto-version compatability for encstrings, huge changes to sockets
Socket userdata will now be a Socket **; we need to allow C/C++ to
allocate the space so we can have constructors in the struct.
Do not memset a socket! It now contains objects that would be damaged.
Starting to map socket activity into a receive queue and event queue...
unsure of how to handle data.

TODO: listen sockets are largely untouched (ie. broken)
Need to update GC routines, other metatable stuff, and events.