Fixed invalid script path when starting MM outside its base path, do not call cleanup on win32 close button callback
Calling Macro::instance()->cleanup() on close button callback may result
in destruction of the stack while the VM is still running. As such, it
will cause a SEGFAULT.

TODO: Hook Lua. On callback, attempt to gracefully terminate the script,
and wait for it to signal when it is OK to continue. Now a cleanup()
should be acceptable.